+++ crates/release_plz_core/src/changelog.rs
@@ -413,7 +413,7 @@ fn default_changelog_config(header: Option<String>) -> ChangelogConfig {
 
 fn default_changelog_body_config() -> &'static str {
     r#"
-## [{{ version | trim_start_matches(pat="v") }}]{%- if release_link -%}({{ release_link }}){% endif %} - {{ timestamp | date(format="%Y-%m-%d") }}
+## [{{ version }}]{%- if release_link -%}({{ release_link }}){% endif %} - {{ timestamp | date(format="%Y-%m-%d") }}
 {% for group, commits in commits | group_by(attribute="group") %}
 ### {{ group | upper_first }}