+++ crates/release_plz_core/src/command/release.rs
@@ -18,12 +18,12 @@ use tracing::{debug, info, instrument, warn};
 use url::Url;
 
 use crate::{
-    CHANGELOG_FILENAME, DEFAULT_BRANCH_PREFIX, GitBackend, PackagePath, Project, ReleaseMetadata,
-    ReleaseMetadataBuilder, Remote,
+    CHANGELOG_FILENAME, DEFAULT_BRANCH_PREFIX, GitForge, PackagePath, Project, Publishable as _,
+    ReleaseMetadata, ReleaseMetadataBuilder, Remote,
     cargo::{CargoIndex, CargoRegistry, CmdOutput, is_published, run_cargo, wait_until_published},
     cargo_hash_kind::get_hash_kind,
     changelog_parser,
-    git::backend::GitClient,
+    git::forge::GitClient,
     pr_parser::{Pr, prs_from_text},
 };
 
@@ -199,6 +199,31 @@ impl ReleaseRequest {
             .or(cargo_utils::registry_token(self.registry.as_deref())?);
         Ok(token)
     }
+
+    /// Checks for inconsistency in the `publish` fields in the workspace metadata and release-plz config.
+    ///
+    /// If there is no inconsistency, returns Ok(())
+    ///
+    /// # Errors
+    ///
+    /// Errors if any package has `publish = false` or `publish = []` in the Cargo.toml
+    /// but has `publish = true` in the release-plz configuration.
+    pub fn check_publish_fields(&self) -> anyhow::Result<()> {
+        let publish_fields = self.packages_config.publish_overrides_fields();
+
+        for package in &self.metadata.packages {
+            if !package.is_publishable() {
+                if let Some(should_publish) = publish_fields.get(&package.name) {
+                    anyhow::ensure!(
+                        !should_publish,
+                        "Package `{}` has `publish = false` or `publish = []` in the Cargo.toml, but it has `publish = true` in the release-plz configuration.",
+                        package.name
+                    );
+                }
+            }
+        }
+        Ok(())
+    }
 }
 
 impl ReleaseMetadataBuilder for ReleaseRequest {
@@ -238,6 +263,19 @@ impl PackagesConfig {
 
     pub fn overridden_packages(&self) -> HashSet<&str> {
         self.overrides.keys().map(|s| s.as_str()).collect()
+    }
+
+    // Return the `publish` fields explicitly set in the
+    // `[[package]]` section of the release-plz config.
+    // I.e. `publish` isn't inherited from the `[workspace]` section of the
+    // release-plz config.
+    pub fn publish_overrides_fields(&self) -> BTreeMap<String, bool> {
+        self.overrides
+            .iter()
+            .map(|(package_name, release_config)| {
+                (package_name.clone(), release_config.publish().is_enabled())
+            })
+            .collect()
     }
 }
 
@@ -470,8 +508,8 @@ impl GitTagConfig {
 
 #[derive(Debug)]
 pub struct GitRelease {
-    /// Kind of Git Backend.
-    pub backend: GitBackend,
+    /// Kind of Git Forge.
+    pub forge: GitForge,
 }
 
 #[derive(Serialize, Default, Debug)]
@@ -661,7 +699,7 @@ async fn should_release(
     }
 }
 
-fn is_pr_commit_in_original_branch(repo: &Repo, commit: &crate::git::backend::PrCommit) -> bool {
+fn is_pr_commit_in_original_branch(repo: &Repo, commit: &crate::git::forge::PrCommit) -> bool {
     let branches_of_commit = repo.get_branches_of_commit(&commit.sha);
     if let Ok(branches) = branches_of_commit {
         branches.contains(&repo.original_branch().to_string())
@@ -893,8 +931,8 @@ fn get_git_client(input: &ReleaseRequest) -> anyhow::Result<GitClient> {
     let git_release = input
         .git_release
         .as_ref()
-        .context("git release not configured. Did you specify git-token and backend?")?;
-    GitClient::new(git_release.backend.clone())
+        .context("git release not configured. Did you specify git-token and forge?")?;
+    GitClient::new(git_release.forge.clone())
 }
 
 #[derive(Debug)]
@@ -1130,5 +1168,20 @@ mod tests {
         // assert the index is inherited from the workspace after the env var
         // is cleared.
         assert_eq!(non_overriden_maybe_registry_index, None);
+    }
+
+    #[test]
+    fn check_publish_fields_works() {
+        // fake_metadata() has `publish = false` in the Cargo.toml
+        let mut request = ReleaseRequest::new(fake_metadata());
+        request = request.with_package_config(
+            "fake_package".to_string(),
+            ReleaseConfig {
+                publish: PublishConfig { enabled: true },
+                ..Default::default()
+            },
+        );
+
+        assert!(request.check_publish_fields().is_err());
     }
 }