+++ crates/release_plz_core/CHANGELOG.md
@@ -7,6 +7,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.34.0](https://github.com/release-plz/release-plz/compare/release_plz_core-v0.33.8...release_plz_core-v0.34.0) - 2025-04-18
+
+### Added
+
+- Alias the `--backend` flag to `--forge` ([#2178](https://github.com/release-plz/release-plz/pull/2178))
+
+### Other
+
+- *(changelog)* stop trimming initial `v` in version ([#2189](https://github.com/release-plz/release-plz/pull/2189))
+
+## [0.33.8](https://github.com/release-plz/release-plz/compare/release_plz_core-v0.33.7...release_plz_core-v0.33.8) - 2025-04-13
+
+### Added
+
+- default "remove_source_branch=true" when creating MR in GitLab ([#2173](https://github.com/release-plz/release-plz/pull/2173))
+
+### Fixed
+
+- fail early if publish = false in Cargo.toml and publish = true in release-plz config ([#2176](https://github.com/release-plz/release-plz/pull/2176))
+
+## [0.33.7](https://github.com/release-plz/release-plz/compare/release_plz_core-v0.33.6...release_plz_core-v0.33.7) - 2025-04-10
+
+### Fixed
+
+- canonicalize paths when comparing package contents ([#2156](https://github.com/release-plz/release-plz/pull/2156))
+- fix `last` clippy lint and ignore security vulnerabilities ([#2166](https://github.com/release-plz/release-plz/pull/2166))
+
 ## [0.33.6](https://github.com/release-plz/release-plz/compare/release_plz_core-v0.33.5...release_plz_core-v0.33.6) - 2025-04-02
 
 ### Fixed